feat: add "docs" section
This adds a new type of post that doesn't have a date associated.
This is for random things I'd like to put up but don't want a time
trail. My intent is to update these over time and it doesn't make sense
to have a date attached.

To facilitate this, I have created a `PreviewList` component that is
intended to allow for a custom preview component to be passed but does
not currently allow to do this.

A small git tip has been added and the navbar now has a "docs" section
containing this type of page.